Move producer helpers to private module.
