+++ src/twisted/internet/test/test_tls.py
@@ -340,13 +340,13 @@ class AbortSSLConnectionTests(ReactorBuilder, AbortConnectionMixin, ContextGener
 
     def buildReactor(self):
         reactor = ReactorBuilder.buildReactor(self)
-        from twisted.internet import producer_helpers
+        from twisted.internet import _producer_helpers
 
         # Patch twisted.protocols.tls to use this reactor, until we get
         # around to fixing #5206, or the TLS code uses an explicit reactor:
         cooperator = Cooperator(
             scheduler=lambda x: reactor.callLater(0.00001, x))
-        self.patch(producer_helpers, "cooperate", cooperator.cooperate)
+        self.patch(_producer_helpers, "cooperate", cooperator.cooperate)
         return reactor
