+++ src/twisted/protocols/tls.py
@@ -58,7 +58,7 @@ from twisted.internet.interfaces import (
     IProtocolNegotiationFactory, IHandshakeListener
 )
 from twisted.internet.main import CONNECTION_LOST
-from twisted.internet.producer_helpers import _PullToPush
+from twisted.internet._producer_helpers import _PullToPush
 from twisted.internet.protocol import Protocol
 from twisted.internet._sslverify import _setAcceptableProtocols
 from twisted.protocols.policies import ProtocolWrapper, WrappingFactory
